Minor refactor for statistics collect job
