+++ kernel/schedule/core/src/test/resources/META-INF/services/org.apache.shardingsphere.infra.metadata.statistics.collector.ShardingSphereStatisticsCollector
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.schedule.core.job.statistics.ShardingSphereStatisticsCollectorFixture
+org.apache.shardingsphere.schedule.core.job.statistics.collect.StatisticsCollectorFixture
